Make bup 0.33 build work

+++ deblib.py
@@ -149,16 +149,18 @@ def wget_download(url):
     # Rename to the directory name the rest of the code expects
     os.rename(prefix, get_name())
 
-def remove_metadata_directories(directory):
+def remove_metadata_directories(directory, keep_git=False, keep_debian=False):
     """
     Remove subdirectories of directory like .git or debian
     """
     # Remove .git & old build directory
-    shutil.rmtree(os.path.join(directory, ".git"), ignore_errors=True)
-    shutil.rmtree(os.path.join(directory, "debian"), ignore_errors=True)
+    if not keep_git:
+        shutil.rmtree(os.path.join(directory, ".git"), ignore_errors=True)
+    if not keep_debian:
+        shutil.rmtree(os.path.join(directory, "debian"), ignore_errors=True)
 
-def pack_source(ext="xz"):
-    remove_metadata_directories(get_name())
+def pack_source(ext="xz", keep_git=False):
+    remove_metadata_directories(get_name(), keep_git=keep_git)
     # Pack source archive
     outfilename = "{}_{}.orig.tar.{}".format(get_name(), get_version(), ext)
     # Delete old archive if it exist
